Move "raisebox" into functions/raisebox.js. (#1331)
